tools/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
GeoAlchemy2==0.15.2
geopandas==1.0.1
greenlet==3.0.3
idna==3.7
lxml==5.2.2
numpy==2.0.1
OWSLib==0.31.0
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
shapely==2.0.5
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
